- reworked addressbook search functions (AB_SearchEntry) completly. Now a   search in the addressbook should perform faster. This will result in a   faster name completion in AddressFields and if a mail is going to be sent   out - fixed some addressbook<->NListtree realtes bugs/issues. - fixed Distribution List bug (BUG #431536)

+++ yamos/YAM_UT.c
@@ -2469,11 +2469,12 @@ ULONG DoSuperNew(struct IClass *cl, Object *obj, ULONG tag1, ...)
 ///
 /// GetMUI
 //  Gets an attribute value from a MUI object
-int GetMUI(Object *obj,int attr)
+LONG GetMUI(Object *obj, ULONG attr)
 {
    LONG b;
-   get(obj,attr,&b);
-   return (int)b;
+
+   get(obj, attr, &b);
+   return b;
 }
 ///
 /// GetMUIStringPtr
